Sort consumer-group related commands output by partition
When users run the describe or reset consumer group commands, the output is jumbled and not in order which makes it hard to comprehend. This code change addresses it by sorting the output by the partition id.

No unit tests exist specifically for the output so I did not create new ones.